Added NPC puppets and idle animations
Wound up not creating a shared skeleton for NPCs; learned the hard way that in Unity, if you want multiple puppets to share the same rig, they have to have the same number of layers. That wasn't the case with this cast of characters. Instead, I wound up rigging and animating each character individually. I'm a little glad that it turned out this way because it allowed me to put more characterization into how the NPCs held themselves-- it gives them a bit of personality.