Built Neighborhood, Woods
Added in most of the assets (need to rework checkered pattern on stall cloth) and started building the environment in the Neighborhood and Woods scenes. It's a little less pastel than I would've liked, but I'm trying to go for something really simplistic, colorful, and childish; might add in some more 2D elements to give it a more cartoony feel if I have the opportunity.