Combat Animations... and Walnut Explosion
We thought it'd be fun to create combat animations; because each character has an individual rig, we weren't able to create a shared set of battle animations that we could apply between every battling character, so we decided to make the best of a bad situation by making the combat animations unique for each character and focus on having their movements emphasize their personality. Meanie's animations are a little floaty and innocent; for example, when she takes damage, she leans back on one foot. Buck's attack relies on his antlers. Because he's a strong, stocky character, his movements are intense and involve a lot of arm motion. Bibbleboo's animations are very timid-- when she takes damage, she flinches, her tail curls between her legs, and her ears pin against her head. Of course, we also finally animated Wally exploding. It's not the most violent thing in the world, but his limbs do go flying. I'd like this game to be bloody, but this is also a relatively unserious death, so we thought it fitting to give him an explosion animation to match.