Built Doenuts Environment
Built the Doenuts environment (a.k.a., the bakery at which Meanie fights Buck and buys the cake that eventually explodes Walnut). I wanted to go for something very nostalgic, cutesy, and picturesque-- not entirely realistic (as per the almost absurd feeling of the game) with pastries displayed on plates and cash being given registerless over a counter, but something a child might imagine a bakery to look like. I stuck with the earthy/gentle green palette I chose for the exterior.